refactor: update credit conversion imports and usage
Refactored credit conversion functions by moving them to a dedicated utils file. Updated import paths in seed and job action files to reflect these changes, ensuring consistent usage of the new utility functions.
