+++ web-app/prisma/seed.ts
@@ -6,7 +6,7 @@ import {
 import crypto from "crypto";
 
 import { getEnvSecrets } from "@/config/env.config";
-import { convertCreditsToBaseUnits } from "@/lib/db/services/credit.service";
+import { convertCreditsToBaseUnits } from "@/lib/db/utils/credit.utils";
 
 import { hashPassword } from "./util/password";
 
@@ -54,7 +54,7 @@ const seedUser = async (): Promise<string> => {
 
   const creditTransaction = await prisma.creditTransaction.create({
     data: {
-      amount: await convertCreditsToBaseUnits(1000.5123),
+      amount: convertCreditsToBaseUnits(1000.5123),
       type: CreditTransactionType.TOP_UP,
       userId: user.id,
       includedFee: 0,
